Extract models into their own package
