+++ database/mockdatabase.go
@@ -1,5 +1,9 @@
 package database
 
+import (
+	"github.com/darthhater/bored-board-service/model"
+)
+
 type MockDatabase struct {
 }
 
@@ -7,8 +11,8 @@ func (m *MockDatabase) InitDb(s string, e string) error {
 	return nil
 }
 
-func (m *MockDatabase) GetThread(threadId string) (Thread, error) {
-	result := Thread{ Id: "", UserId: "admin", Title: "What the heck", PostedAt: "A time" }
+func (m *MockDatabase) GetThread(threadId string) (model.Thread, error) {
+	result := model.Thread{ Id: "", UserId: "admin", Title: "What the heck", PostedAt: "A time" }
 
 	return result, nil
 }
