+++ model/post.go
@@ -1,0 +1,9 @@
+package model
+
+type Post struct {
+	Id string
+	ThreadId string
+	UserId string
+	Body string
+	PostedAt string
+}
