+++ model/thread.go
@@ -1,0 +1,8 @@
+package model
+
+type Thread struct {
+	Id string
+	UserId string
+	Title string
+	PostedAt string
+}
